[plugin] unset PYTHONPATH before starting Kodi
PYTHONPATH environment variable is set by enigma2 to /usr/lib/enigma2/python.
We are starting Kodi using same environment variables as enigma2. When kodi
starts python interpreter, it uses PYTHONPATH which means that sys.path
has PYTHONPATH set by enigma2 in it.

This causes problems to addons which import modules with the same
name as in enigma2 python directory.

For example Navi-X plugin tries to import skin module, which also
exists in enigma2 python directory. Since this directory is closer
to the beginning of the sys.path's list, it's searched first for skin
module which is then selected. Addon execution is then terminated with
error, since wrong module was imported.

To avoid this problem we unset PYTHONPATH environment variable before starting
Kodi.

thx mx3l